Remove dependency on derivative crate
The derivative crate is a heavyweight dependency, and one that is
relatively unmaintained: its last release was eight months ago, and it
is one of the few crates in the ecosystem that still depends upon
pre-1.0 syn, quote, and proc-macro2.

In the interest of keeping tokio-serde lightweight, and in the spirit of
[0], I think it makes more sense to simply add the correct Default
implementations by hand, rather than pulling in all of derivative.

[0]: https://raphlinus.github.io/rust/2019/08/21/rust-bloat.html